backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
openai==1.59.5
python-docx==1.1.2
PyPDF2==3.0.1
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
pydantic==2.10.4
